requirements.txt
fastapi
uvicorn
kiteconnect
google-cloud-firestore
jinja2
pandas
pandas-ta-classic